requirements.txt
# Django и async поддержка
Django==5.1.5
asgiref==3.8.1
psycopg[binary]==3.2.3
python-decouple==3.8
uvicorn==0.41.0

# Telegram бот
aiogram==3.16.0
aiohttp==3.10.11

# Биржа и торговля
ccxt==4.4.47
numpy==2.2.1
pandas==2.2.3
ta==0.11.0

# Celery и асинхронные задачи
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Шифрование
cryptography==44.0.0

# Утилиты
python-dotenv==1.0.1
